easy implementation to other projects added. Created UnityCommunicatorClient() class. Added comments.
