+++ TCPGeometrics/Assets/Scripts/UCGameObjectController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1a795c05eee7a4f2f84cbf4f9ac915b3
+guid: 5f2d9dfd50d9046f180af504e295be34
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
