+++ TCPGeometrics/Assets/Scripts/UCJSONObjectTemplates.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f11312bd61ca4ef4954a97fc445abb8
+guid: 390ecfd03401848bb895a5b1d776f75a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
