+++ TCPGeometrics/Assets/Scripts/UCManageCapture.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b939d835b5824385b9b363030000e21
+guid: 3a02b41546cdb43ab83e848d87290fd0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
