--- TCPGeometrics/Assets/StreamingAssets/Encoded.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 389e7714dc24642c68d66f2a8a74eb0b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
