--- TCPGeometrics/Assets/StreamingAssets/parameters_geometrics1.json
@@ -1,65 +1,0 @@
-{
-  "message": "Paramter JSON File number 1, hello world",
-  "sceneID": 1,
-  "Object0": {
-    "position": [-9,12.65,-9],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [45,45,0],
-    "orthographic": 10,
-    "fieldOfView": 5
-  },
-  "Object1": {
-    "position": [1,5,1],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [100,0,0],
-    "color":[1.0,1.0,1.0],
-    "intensity":0.48,
-    "active": 1,
-    "range": 72.9,
-    "spotAngle": 64
-  },
-  "Object2":  {
-    "position": [-0.77,5.2,-2.38],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [0,0,0],
-    "color":[1.0,1.0,1.0],
-    "intensity":9.9,
-    "active": 1,
-    "range": 10
-  },
-  "Object3":{
-    "position": [0,0,0],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [50,-30,0],
-    "color":[1.0,1.0,1.0],
-    "intensity":0.31,
-    "active": 1
-  },
-  "Object4": {
-    "position": [1,2,0],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [0,45,45],
-    "color":[1,1,1],
-    "transparency": 1,
-    "metallic": 0,
-    "glossiness": 0
-  },
-  "Object5": {
-    "position": [2,1,1],
-    "scale": [3,3,3],
-    "rotation": [0,0,0],
-    "color":[1,1,1],
-    "transparency": 1,
-    "metallic": 0,
-    "glossiness": 0
-  },
-  "Object6": {
-    "position": [-2,1,-1],
-    "scale": [1.0,1.0,1.0],
-    "rotation": [0,0,0],
-    "color":[1,1,1],
-    "transparency": 1,
-    "metallic": 0,
-    "glossiness": 0.5
-  }
-}
