--- TCPGeometrics/Assets/StreamingAssets/parameters_geometrics1.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c3b7b716787584440b2585c8655aa348
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
