--- TCPGeometrics/Assets/StreamingAssets/tcpconfig.json.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2e5892c60232b43cea6cbbbcc5c47cdd
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
